requirements.txt
python-dotenv
openai
pydantic
pdfplumber
python-docx
supabase
logfire[openai]
exa-py